Jwt auth | Deleted spec tests | Auth

+++ src/hire/hire.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpStatus, HttpException } from '@nestjs/common';
 import { CreateHireDto } from './dto/create-hire.dto';
 import { UpdateHireDto } from './dto/update-hire.dto';
 import { Hire } from './hire.entity';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, createQueryBuilder } from 'typeorm';
+import { Repository } from 'typeorm';
+
 
 @Injectable()
 export class HireService {
@@ -13,7 +14,6 @@ export class HireService {
     ) {}
 
     async create(createHireDto: CreateHireDto) {
-
         const hire = await this.hireRepository.query(`
             select
             "hire.carId",
@@ -23,11 +23,18 @@ export class HireService {
             "hire.endRateDate"::date - "hire.createRateDate"::date as days,
             (select days * rate.price)
             from hire, rate
-            `)
+            `);
         hire.car = createHireDto.car;
         hire.rate = createHireDto.rate;
-        hire.createRateDate = createHireDto.createRateDate
-        hire.endRateDate = createHireDto.endRateDate
+        hire.createRateDate = createHireDto.createRateDate;
+        hire.endRateDate = createHireDto.endRateDate;
+
+        if (hire.days > 30) {
+            throw new HttpException(
+                'Сar rental no more than 30 days',
+                HttpStatus.FORBIDDEN,
+            );
+        }
         await this.hireRepository.save(hire);
         return hire;
     }
@@ -37,14 +44,27 @@ export class HireService {
     }
 
     async findOne(id: number): Promise<Hire> {
-        return this.hireRepository.findOne(id);
+        const hire = await this.hireRepository.findOne(id);
+        if (hire) {
+            return hire;
+        }
+        throw new HttpException('Hire not found', HttpStatus.NOT_FOUND);
     }
 
     async update(id: number, updateHireDto: UpdateHireDto) {
-        return this.hireRepository.update(id, updateHireDto);
+        await this.hireRepository.update(id, updateHireDto);
+        const updateHire = await this.hireRepository.findOne(id);
+        if (updateHire) {
+            updateHire;
+        }
+        throw new HttpException('Hire not found', HttpStatus.NOT_FOUND);
     }
 
-    async remove(id: number): Promise<void> {
+    async remove(id: number) {
+        const removeHire = await this.hireRepository.delete(id);
+        if (!removeHire.affected) {
+            throw new HttpException('Hire not found', HttpStatus.NOT_FOUND);
+        }
         await this.hireRepository.delete(id);
     }
 }
